style: context 분리
