+++ pages/bookcase/index.tsx
@@ -2,7 +2,7 @@ import Cards from "../../components/bookcase/Cards";
 import Navigation from "../../components/bookcase/Navigation";
 import { MainLayout } from "../../components/layout";
 import { MainHeader } from "../../components/main";
-import { BookcaseNavigationProvider } from "../../util/hooks/useBookcase";
+import { BookcaseNavigationProvider } from "../../util/context/BookcaseProvider";
 import useUser from "../../util/hooks/useUser";
 
 export default function Bookcase() {
